Enhancing Cognitive Disorder Prediction Through Multi-Modal Neuroimaging Fusion.
The convergence of frontier technologies with the applied science of medicine has created an environment in healthcare research that has stood on the brink of revolutionary discoveries. One of the most exciting and rapidly developing areas is that of early prediction and understanding of cognitive disorders. This is going to keep the limits of current research expanding. Responding to this call of action, the present dissertation picks up where the last left off and starts a journey of exploration with a title: Advancing Cognitive Disorder Prediction Through Multi-Modal Neuroimaging Fusion. This work unites state-of-the-art advances in neuroimaging with frontier artificial intelligence techniques in the state of the art using a multi-modal approach to elicit insights from a variety of neuroimaging techniques, including MRI, PET, and fMRI. Such integration has the potential to provide a holistic view of brain functioning and better predictive models in the complex landscape of cognitive disorders.